1.x: javac- and javadoc-related cleanup in components, part 1

+++ src/test/java/rx/internal/operators/OperatorTakeUntilPredicateTest.java
@@ -147,7 +147,7 @@ public class OperatorTakeUntilPredicateTest {
         }).subscribe(ts);
         ts.assertTerminalEvent();
         ts.assertNotCompleted();
-        assertEquals(1, (int) ts.getOnErrorEvents().size());
+        assertEquals(1, ts.getOnErrorEvents().size());
         assertTrue(ts.getOnErrorEvents().get(0).getCause().getMessage().contains("abc"));
     }
 }
